moved a few structures from `zstd_internal.h` to `zstd_compress.h`
which is a more precise scope
